Moved defined browser initialization to separate classes
